requirements.txt

pyenchant
streamlit
